Create README.md
This project is developed in such a way that an user can select the option according to their choice. For Ex: If he wants lo learn HTML Course, If he choose the Learn HTML it will provide the course details of HTML Course. This will students for better understanding of their choice.